update folder and rename files
